+++ app/routes/[robots.txt].tsx
@@ -36,7 +36,7 @@ export const loader: LoaderFunction = () => {
   Allow: /wow/pet-marketshare
   Allow: /wow/weekly-price-group-delta
   Allow: /ffxiv/scrip-exchange
-  Allow: /ffxiv/fc-seals-exchange
+  Allow: /ffxiv/gc-seals-exchange
   Allow: /ffxiv/extended-history
   Allow: /ffxiv/best-deals/recommended
   Allow: /ffxiv/weekly-price-group-delta
